#741 small refactoring to update directory structure, based on PR comments
