+++ src/hgvs/pretty/console/tx_alig_renderer.py
@@ -1,7 +1,7 @@
 import math
 
 from hgvs.pretty.models import VariantData
-from hgvs.pretty.renderer.renderer import BasicRenderer
+from hgvs.pretty.console.renderer import BasicRenderer
 
 
 class TxAligRenderer(BasicRenderer):
@@ -17,7 +17,7 @@ class TxAligRenderer(BasicRenderer):
         if not data.var_c_or_n:
             return ""
 
-        from hgvs.pretty_print import ENDC, TPURPLE, TYELLOW
+        from hgvs.pretty.pretty_print import ENDC, TPURPLE, TYELLOW
 
         var_str = ""
